fix(prompt): each Executor has its own settings, 'a little copying is better than a little dependency'
